Ports frogs from my /tg/ PR (#48)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
## Original PR: https://github.com/tgstation/tgstation/pull/78537

## About The Pull Request
It has been over a year since we have got a new pAI holoform, so I
present to you, the fwog.

![image](https://github.com/Towelstation13/towelstation13/assets/20053168/e86cc27a-9a67-4b3e-9ba2-a102885fd767)
More importantly, it can be worn on the head like a normal frog.

![image](https://github.com/Towelstation13/towelstation13/assets/20053168/a91fc236-eade-4738-a6d9-0021a0039d77)

All of the new icons

![image](https://github.com/Towelstation13/towelstation13/assets/20053168/b7ad19fc-941f-4a10-bf14-0632eefcae99)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## How This Contributes To The Towelstation Roleplay Experience
More holoform variety for pAIs is nice, especially one that can be worn
as a hat.

<!-- Please add a short description of why you think these changes would
benefit the game and the roleplay atmosphere of the server. If you can't
justify it in words, it might not be worth adding. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
image: adds a frog holoform for pAIs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->