(MIRROR) Ports a port of a port | Ports paper masks (#61)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Mirror of https://github.com/Skyrat-SS13/Skyrat-tg/pull/24180 and
https://github.com/Skyrat-SS13/Skyrat-tg/pull/24326

![image](https://github.com/Towelstation13/towelstation13/assets/20053168/711a8609-6e7c-48c4-9c24-be03ac38c25c)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## How This Contributes To The Towelstation Roleplay Experience
More customization
<!-- Please add a short description of why you think these changes would
benefit the game and the roleplay atmosphere of the server. If you can't
justify it in words, it might not be worth adding. -->

## Proof of Testing

<!-- Include any screenshots/videos/debugging steps of the code
functioning successfully, between the </summary> and </details> code
blocks. -->
<!-- To our mappers and spriters: Posting screenshots of content INSIDE
EDITORS (aseprite, PDN, SDMM, ect) is NOT valid proof of testing. Please
make sure that you COMPILE the game and provide PROOF you tested your
edits. -->

<details>
<summary>Screenshots/Videos</summary>
  
</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Nerev4r, vinylspiders
image: The crew's knowledge of origami and papercrafting has been
extended to making paper masks. Find them in the loadout, or just make
them!
fix: fixed an issue that could cause the hair the layering option to
reset itself upon unequipping a paper mask
qol: paper masks have an action button for adjusting the mask drawing +
hair layering, as well as the ability to hide the strap with ctrl click.
changing the drawings on the mask now require a pen.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->